CardType and Points properties added to Client class

+++ Micromania.Console/Game.cs
@@ -9,14 +9,14 @@ namespace Micromania.Console
     public class Game : Entity
     {
         public virtual string Title { get; protected set; }
-        public virtual decimal Price { get; protected set; }
+        public virtual double Price { get; protected set; }
         public virtual int Points { get; protected set; }
 
         public Game()
         {
         }
 
-        public Game(string title, decimal price)
+        public Game(string title, double price)
         {
             Title = title;
             Price = price;
